solutions/01-manual/python-year/requirements.txt
anyio==3.6.1
asgiref==3.5.2
backoff==1.10.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
fastapi==0.80.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
h11==0.13.0
idna==3.3
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-asgi==0.36b0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-sdk==1.15.0
protobuf==3.15.0
pydantic==1.9.2
requests==2.28.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wrapt==1.14.1
